backend/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
httpx>=0.26.0
feedparser>=6.0.12
python-multipart>=0.0.9
pydantic>=2.0.0
python-dateutil>=2.9.0
beautifulsoup4>=4.12.0
lxml>=4.9.0

